src/requirements.txt
<<<<<<< HEAD
amqp==2.5.2
asgiref==3.2.7
autopep8==1.5.2
=======
asgiref==3.2.7
>>>>>>> a8aa4e4ba879bca99be5dfe897736f99e6750c8e
backcall==0.1.0
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
Django==3.0.6
django-debug-toolbar==2.2
django-extensions==2.2.9
django-silk==4.0.1
entrypoints==0.3
Faker==4.0.3
flake8==3.7.9
flake8-builtins==1.5.2
flake8-import-order==0.18.1
flake8-print==3.1.4
gprof2dot==2019.11.30
idna==2.9
importlib-metadata==1.6.0
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
kombu==4.6.8
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.9
vine==1.3.0
wcwidth==0.1.9
zipp==3.1.0
